+++ scripts/lint.sh
@@ -11,4 +11,24 @@ if rustup component add clippy; then
     `# Exit with a nonzero code if there are clippy warnings` \
     -Dwarnings \
     "$@"
+  cargo clippy --manifest-path=opentelemetry-jaeger/Cargo.toml --all-targets --features "surf_collector_client" --no-default-features -- \
+    `# Exit with a nonzero code if there are clippy warnings` \
+    -Dwarnings \
+    "$@"
+  cargo clippy --manifest-path=opentelemetry-jaeger/Cargo.toml --all-targets --features "isahc_collector_client" --no-default-features -- \
+    `# Exit with a nonzero code if there are clippy warnings` \
+    -Dwarnings \
+    "$@"
+  cargo clippy --manifest-path=opentelemetry-jaeger/Cargo.toml --all-targets --features "reqwest_blocking_collector_client" --no-default-features -- \
+    `# Exit with a nonzero code if there are clippy warnings` \
+    -Dwarnings \
+    "$@"
+  cargo clippy --manifest-path=opentelemetry-jaeger/Cargo.toml --all-targets --features "reqwest_collector_client" --no-default-features -- \
+    `# Exit with a nonzero code if there are clippy warnings` \
+    -Dwarnings \
+    "$@"
+  cargo clippy --manifest-path=opentelemetry-jaeger/Cargo.toml --all-targets --features "collector_client" --no-default-features -- \
+    `# Exit with a nonzero code if there are clippy warnings` \
+    -Dwarnings \
+    "$@"
 fi